check_kernel_printf.c: Handle 'fmt ? : "default"'
It is not completely unlikely that someone could make a call like
sprintf(buf, fmt ? : "some default");. In that case, ->cond_true is
NULL, and we should therefore use ->conditional instead. That is very
unlikely to be a string literal which we can actually check (why would
the ?: then be used), but it might come from some macro expansion, and
it's certainly better than segfaulting.