improve trace_params.pl
smatch_warns.txt is over 8 GB for an allmodconfig kernel. That takes a
lot of time to process. We can easily tweak the trace_params.pl script
to allow a list of function/parameter index pairs, instead of just one
pair. So do that and update the gen_* scripts accordingly.

On my machine, this cuts the time for running
smatch_scripts/gen_gfp_flags.sh from 64 to 24 seconds; this is close
to a factor of 3, so indeed most of the time was spent in load_all().

While at it: remove some redundant tmp files, update comment in
check_param_mapper.c, and eliminate false positives coming from the
use of regexp match instead of string equality. For example,
"devm_kzalloc 1" used to be in the output of the gfp script, but
argument 1 of devm_kzalloc is certainly not a gfp_t.