fill_db_type_size.pl: simplify text_to_int
Profiling shows this sub at the top of the list at 18% of the total
run-time. Using a lookup table drops that down to 7%.

Not sure whether the value for u64max is a typo or whether there's
some magic reason it needs to be 2**62-1, but I'll leave it for now.

Signed-off-by: Rasmus Villemoes <rv@rasmusvillemoes.dk>